+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add `insert` and `extract` instructions.
 
+### Changed
+- Change type of delta and epsilon time steps to `usize`.
+
 ### Fixed
 - Fix representation of times in assembly; always print seconds. ([#36](https://github.com/fabianschuiki/llhd/issues/36))